# docs/swagger.yaml
basePath: /
definitions:
...
  dto.GetMediaRequest:
    properties:
      media_id:
        example: media_123456789
        type: string
    required:
    - media_id
    type: object
...
  dto.MediaData:
    properties:
      action:
        example: upload
        type: string
      download:
        $ref: '#/definitions/dto.MediaDownloadInfo'
      media_id:
        example: media_123456789
        type: string
      media_info:
        $ref: '#/definitions/dto.MediaInfo'
      progress:
        $ref: '#/definitions/dto.MediaUploadProgress'
      status:
        example: success
        type: string
      timestamp:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  dto.MediaDownloadInfo:
    properties:
      download_url:
        example: https://storage.example.com/media_123456789
        type: string
      expires_at:
        example: "2023-01-01T01:00:00Z"
        type: string
      media_id:
        example: media_123456789
        type: string
    type: object
...
  dto.MediaErrorResponse:
    properties:
      code:
        example: INVALID_MEDIA_ID
        type: string
      details:
        example: Media ID must be alphanumeric
        type: string
      message:
        example: Invalid media ID format
        type: string
    type: object
  dto.MediaInfo:
    properties:
      file_name:
        example: image.jpg
        type: string
      media_id:
        example: media_123456789
        type: string
      media_type:
        example: image
        type: string
      mime_type:
        example: image/jpeg
        type: string
      session_id:
        example: default
        type: string
      size:
        example: 1024000
        type: integer
      status:
        example: ready
        type: string
      uploaded_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  dto.MediaResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/dto.MediaData'
      error:
        $ref: '#/definitions/dto.MediaErrorResponse'
      success:
        type: boolean
    type: object
  dto.MediaUploadProgress:
    properties:
      bytes_uploaded:
        example: 768000
        type: integer
      media_id:
        example: media_123456789
        type: string
      progress:
        example: 75.5
        type: number
      status:
        example: uploading
        type: string
      total_bytes:
        example: 1024000
        type: integer
    type: object
...
  dto.UploadMediaRequest:
    properties:
      file_name:
        example: image.jpg
        type: string
      media_type:
        example: image
        type: string
      session_id:
        example: default
        type: string
    required:
    - media_type
    - session_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@meow.com
    name: meow API Support
    url: https://github.com/your-username/meow
  description: A meow API server built with Go, inspired by meow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: meow meow API
  version: "1.0"
paths:
...
  /session/{sessionId}/media/{mediaId}:
    delete:
      consumes:
      - application/json
      description: Delete a media file from the server
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media ID
        in: path
        name: mediaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete media file
      tags:
      - Media
  /session/{sessionId}/media/{mediaId}/compress:
    post:
      consumes:
      - application/json
      description: Compress a media file to reduce its size
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media ID
        in: path
        name: mediaId
        required: true
        type: string
      - description: 'Compression quality (1-100, default: 80)'
        in: query
        name: quality
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Compress media file
      tags:
      - Media
  /session/{sessionId}/media/{mediaId}/convert:
    post:
      consumes:
      - application/json
      description: Convert a media file to a different format
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media ID
        in: path
        name: mediaId
        required: true
        type: string
      - description: Target format (e.g., jpg, png, mp4, mp3)
        in: query
        name: format
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Convert media format
      tags:
      - Media
  /session/{sessionId}/media/{mediaId}/download:
    get:
      consumes:
      - application/json
      description: Download a media file from the server
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media ID
        in: path
        name: mediaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Download media file
      tags:
      - Media
  /session/{sessionId}/media/{mediaId}/metadata:
    get:
      consumes:
      - application/json
      description: Get detailed metadata information about a media file
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media ID
        in: path
        name: mediaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Get media metadata
      tags:
      - Media
  /session/{sessionId}/media/{mediaId}/progress:
    get:
      consumes:
      - application/json
      description: Get the upload progress of a media file
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Media ID
        in: path
        name: mediaId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Get media upload progress
      tags:
      - Media
  /session/{sessionId}/media/info:
    post:
      consumes:
      - application/json
      description: Get information about a specific media file
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Get media request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Get media information
      tags:
      - Media
  /session/{sessionId}/media/list:
    get:
      consumes:
      - application/json
      description: Get a list of all media files for a session
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: 'Limit number of results (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Offset for pagination (default: 0)'
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: List media files
      tags:
      - Media
  /session/{sessionId}/media/upload:
    post:
      consumes:
      - application/json
      description: Upload a media file to the server
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Upload media request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UploadMediaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.MediaResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.MediaResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload media file
      tags:
      - Media
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key authentication. Simply provide your API key directly: "YOUR_API_KEY".
      The system automatically detects if it''s a Global API Key (can access all sessions
      and session management) or a Session-specific API Key (can only access the specific
      session it belongs to).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...