# internal/interfaces/http/handlers/media.go
package handlers

import (
	"net/http"

	"zpmeow/internal/application"
	"zpmeow/internal/infra/wmeow"
	"zpmeow/internal/interfaces/dto"

	"github.com/gin-gonic/gin"
)

type MediaHandler struct {
	sessionService *application.SessionApp
	wmeowService   wmeow.WameowService
}

func NewMediaHandler(sessionService *application.SessionApp, wmeowService wmeow.WameowService) *MediaHandler {
	return &MediaHandler{
		sessionService: sessionService,
		wmeowService:   wmeowService,
	}
}

// @Summary		Upload media file
// @Description	Upload a media file to the server
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string					true	"Session ID"
// @Param			request		body		dto.UploadMediaRequest	true	"Upload media request"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/upload [post]
func (h *MediaHandler) UploadMedia(c *gin.Context) {
	var req dto.UploadMediaRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, dto.NewMediaErrorResponse(
			http.StatusBadRequest,
			"INVALID_REQUEST",
			"Invalid request format",
			err.Error(),
		))
		return
	}

	c.JSON(http.StatusOK, dto.NewMediaSuccessResponse("", "upload", nil))
}

// @Summary		Get media information
// @Description	Get information about a specific media file
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string				true	"Session ID"
// @Param			request		body		dto.GetMediaRequest	true	"Get media request"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		404			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/info [post]
func (h *MediaHandler) GetMedia(c *gin.Context) {
	var req dto.GetMediaRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, dto.NewMediaErrorResponse(
			http.StatusBadRequest,
			"INVALID_REQUEST",
			"Invalid request format",
			err.Error(),
		))
		return
	}

	c.JSON(http.StatusOK, dto.NewMediaSuccessResponse(req.MediaID, "get", nil))
}

// @Summary		Download media file
// @Description	Download a media file from the server
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string	true	"Session ID"
// @Param			mediaId		path		string	true	"Media ID"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		404			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/{mediaId}/download [get]
func (h *MediaHandler) DownloadMedia(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"message": "Download media endpoint - implementation pending",
	})
}

// @Summary		Delete media file
// @Description	Delete a media file from the server
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string	true	"Session ID"
// @Param			mediaId		path		string	true	"Media ID"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		404			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/{mediaId} [delete]
func (h *MediaHandler) DeleteMedia(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"message": "Delete media endpoint - implementation pending",
	})
}

// @Summary		List media files
// @Description	Get a list of all media files for a session
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string	true	"Session ID"
// @Param			limit		query		int		false	"Limit number of results (default: 50)"
// @Param			offset		query		int		false	"Offset for pagination (default: 0)"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/list [get]
func (h *MediaHandler) ListMedia(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"message": "List media endpoint - implementation pending",
	})
}

// @Summary		Get media upload progress
// @Description	Get the upload progress of a media file
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string	true	"Session ID"
// @Param			mediaId		path		string	true	"Media ID"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		404			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/{mediaId}/progress [get]
func (h *MediaHandler) GetMediaProgress(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"message": "Get media progress endpoint - implementation pending",
	})
}

// @Summary		Convert media format
// @Description	Convert a media file to a different format
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string	true	"Session ID"
// @Param			mediaId		path		string	true	"Media ID"
// @Param			format		query		string	true	"Target format (e.g., jpg, png, mp4, mp3)"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		404			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/{mediaId}/convert [post]
func (h *MediaHandler) ConvertMedia(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"message": "Convert media endpoint - implementation pending",
	})
}

// @Summary		Compress media file
// @Description	Compress a media file to reduce its size
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string	true	"Session ID"
// @Param			mediaId		path		string	true	"Media ID"
// @Param			quality		query		int		false	"Compression quality (1-100, default: 80)"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		404			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/{mediaId}/compress [post]
func (h *MediaHandler) CompressMedia(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"message": "Compress media endpoint - implementation pending",
	})
}

// @Summary		Get media metadata
// @Description	Get detailed metadata information about a media file
// @Tags			Media
// @Accept			json
// @Produce		json
// @Param			sessionId	path		string	true	"Session ID"
// @Param			mediaId		path		string	true	"Media ID"
// @Success		200			{object}	dto.MediaResponse
// @Failure		400			{object}	dto.MediaResponse
// @Failure		404			{object}	dto.MediaResponse
// @Failure		500			{object}	dto.MediaResponse
// @Security		ApiKeyAuth
// @Router			/session/{sessionId}/media/{mediaId}/metadata [get]
func (h *MediaHandler) GetMediaMetadata(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{
		"success": true,
		"message": "Get media metadata endpoint - implementation pending",
	})
}